Support password required when changing email
Changing email now requires the user to enter their password first. This
affected the user update route because a new field was created to
transport the password given when a email is updated. The password is
thrown away as soon as it is validated to match the user's password.

The auth method for user profile update was also modified to handle
checking the password given on email update.