Support for meta tags for user profiles
Adds support for inject user details into the public/index.html page via
handlebars to populate the meta tags for social media sites. Currently,
the username is used for the title, the signature is used for the
description, and the user's post count and creation date are used for
any data fields.

This was made possible by adding a new route for robots to consume that
loads only the information mentioned above. For regular users, the same
process is applied but once site loads, they pull down the same
information again.