1.0.6: User Preferences
This commit adds the ability to track a user's preferences across the
app. The items being tracked are posts per page, threads per page, and
the category visible on the boards view.

The update and userByUsername db method were updated to handle saving
the new preference information, and to pull that information back out
respectively. The update method also handles creating a new row in the
users.preferences table if it does not already exist.

There is a small note in terms of storing the collapsed_categories
field. There's a bug in node-postgres that does not handle saving json
arrays well so the array itself in stored as in the cats property of a
larger object. This packaging and unpacking should always be handled at
the DB level so as to seem as a regular json array throughout the rest
of the app.

A preferences db method was also created to just retrieve the
preferences data and not the rest of the user data. It also handles
unpacking the collapsed_categories object.

The find and update routes were also updated to handle the change in
payload (new preference data). An additional preferences route was also
created to pull down just the preference data. This is used when the
frontend app is bootstrapped. This will ensure that the preferences data
is in used at the beginning of page load. The find route is mainly used
for the profile page, and the update route is used whenever the
preference data needs to be updated.